fix: partial fix of bugs that use conditionally independent evidence
The program now runs for all input cases, and no longer crashes when irrelevant evidence is provided. However, in (mostly diagnostic) cases, where indirect evidence is provided, the calculated result is wrong. This seems to be an issue related to (1) the pruneOrder method in VEWE and (2) the joinMarginalize method. Needs more investigation!